moved spotify refresh logic into context to make things more modularized
