+++ components/Spaces/CallTabs/Music/index.js
@@ -1,10 +1,6 @@
-import { useState, useEffect } from 'react';
+import { useState } from 'react';
 import PropTypes from 'prop-types';
-import jwt from 'jsonwebtoken';
 import classNames from 'classnames';
-import axios from 'axios';
-import { useRecoilState } from 'recoil';
-import { differenceInMilliseconds, addMilliseconds } from 'date-fns';
 import { TabList, Tab, Tabs, TabPanel } from 'react-tabs';
 import Button from '@material-ui/core/Button';
 import Typography from '@material-ui/core/Typography';
@@ -12,8 +8,6 @@ import { makeStyles } from '@material-ui/core/styles';
 
 import { useSpotifyContext, ENUM_AUTHENTICATION } from '@/context/spaces/SpotifyContext';
 import renderComponent from '@/utils/renderComponent';
-import getCookie from '@/utils/getCookie';
-import * as spotifyState from '@/atoms/spotify';
 import SpotifySignUpBlocker from './SpotifySignUpBlocker';
 import Player from './Player';
 import HomeTab from './Tabs/Home';
@@ -54,39 +48,8 @@ const useStyles = makeStyles((theme) => ({
 
 const Music = ({ tabs }) => {
   const classes = useStyles();
-  const { getAccessTokenFromCookies, authenticated } = useSpotifyContext();
   const [tabIndex, setTabIndex] = useState(0);
-  const [spotifyRefresh, setSpotifyRefresh] = useRecoilState(spotifyState.refresh);
-
-  useEffect(() => {
-    const spotifySessionJWT = getCookie(document.cookie, 'spotify_session');
-    if (!spotifySessionJWT) return;
-    const spotifySession = jwt.decode(spotifySessionJWT);
-    if (!spotifySession) return;
-    let timeoutDuration = 100; // set to 100 for instance refresh if recoil state is not set
-    if (spotifyRefresh?.refreshDate) {
-      const curDate = new Date();
-      const timeToRefresh = differenceInMilliseconds(spotifyRefresh?.refreshDate, curDate) ?? 3600 * 1000;
-      timeoutDuration = timeToRefresh;
-    }
-    const timeout = setTimeout(() => {
-      if (!spotifySession?.refreshToken) return;
-      axios
-        .post('/api/spotify/refresh-token', { refreshToken: spotifySession.refreshToken })
-        .then((res) => {
-          const expiresIn = res.data.data.expiresIn * 1000;
-          const date = new Date();
-          const expireDate = addMilliseconds(date, expiresIn);
-          const refreshDate = addMilliseconds(date, expiresIn / 4);
-          setSpotifyRefresh({ expiresIn, expireDate, refreshDate });
-          console.debug('Successfully refreshed spotify token:', res.data);
-          getAccessTokenFromCookies();
-        })
-        .catch((err) => console.debug(err));
-    }, timeoutDuration);
-
-    return () => clearTimeout(timeout);
-  }, [spotifyRefresh]);
+  const { authenticated } = useSpotifyContext();
 
   if (authenticated !== ENUM_AUTHENTICATION.AUTHENTICATED) {
     return <SpotifySignUpBlocker loading={authenticated === ENUM_AUTHENTICATION.LOADING} />;
